Dear, Adam
Because I only had fifteen minutes and a headache, I sloppily put together an example LinearOpMode, only for your learning purposes.  At the bottom of the class should be comments that vaguely explain the LinearOpMode class and a couple resources for more info.  Again, none of the code is properly written, but they're only there to express ideas.

Joel